user-service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pymongo==4.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
requests==2.32.3